requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
ecdsa==0.19.1
Flask==3.1.2
Flask-Caching==2.3.1
Flask-Limiter==4.1.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
marshmallow==4.2.0
marshmallow-sqlalchemy==1.4.2
mysql-connector-python==9.5.0
ordered-set==4.1.0
packaging==25.0
pyasn1==0.6.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.5
wrapt==2.0.1
